orchestrator/requirements.txt
quart==0.20.1
uvicorn==0.34.0
redis[hiredis]==5.0.3
msgspec==0.18.6
grpcio==1.78.0
protobuf>=6.31.1
circuitbreaker==2.1.3
